% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data1}
\alias{data1}
\title{Example data from Charlie's webpage}
\format{An object of class \code{data.frame} with 30 rows and 2 columns.}
\usage{
data1
}
\description{
Example data from Charlie's webpage
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p4.txt} #data1

\url{http://www.stat.umn.edu/geyer/3701/data/q1p7c.txt} #data2

\url{http://www.stat.umn.edu/geyer/3701/data/q2p3.rda} #data3

\url{http://www.stat.umn.edu/geyer/3701/data/q2p1.rda} #data4
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
