% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes-matcalc-grapes.r
\name{\%matcalc\%}
\alias{\%matcalc\%}
\title{Matrix Calculator for X^T A^{-1} X}
\usage{
A \%matcalc\% X
}
\arguments{
\item{A}{matrix (must be invertible)}

\item{X}{numerical vector}
}
\value{
calculated matrix
}
\description{
Working on a matrix, A, and a numerical vector X, performs the calculation in the title line.
}
\examples{


data4[[1]] `\%matcalc\%` data4[[2]]
}
