% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MercurioRyanTools-package.r
\docType{package}
\name{MercurioRyanTools}
\alias{MercurioRyanTools}
\alias{MercurioRyanTools-package}
\title{MercurioRyanTools.}
\description{
3/7/2018
This package is being submited as HW#3 for STAT 3701 at UMN.
}
\details{
I have not included the Quiz/HW functions that I have deemed redundant.
}
