% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstandardize.R
\name{mstandardize}
\alias{mstandardize}
\title{Matrix Standardization}
\usage{
mstandardize(M)
}
\arguments{
\item{M}{matrix with multiple rows.}
}
\value{
A new, standardized matrix.
}
\description{
Standardize the columns of a matrix.
}
\examples{

mstandardize(data3)

}
