% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myapply.R
\name{myapply}
\alias{myapply}
\title{myapply}
\usage{
myapply(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{matrix}

\item{MARGIN}{is either 1 or 2, for rows or columns}

\item{FUN}{is some function to be applied.}
}
\value{
Answer in simplest form - matrix or array.
}
\description{
Simplified version of apply for STAT 3701
}
\examples{

fred <- matrix(1:6, nrow = 3, ncol = 2)

myapply (fred, 2, sum)
myapply (fred, 1, sum)

}
