% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylogl.r
\name{mylogl}
\alias{mylogl}
\title{Log Likelihood Estimator}
\usage{
mylogl(x, func, interval)
}
\arguments{
\item{x}{data vector}

\item{func}{function with arguments theta (the parameter) and x (the data vector)}

\item{interval}{search range(numerical vector of length 2)}
}
\value{
The maximum likelihood estimator for the function: one numeric value
}
\description{
Using a data vector and its approximate distribution, find the maximum log-likelihood estimator for a given interval.
}
\details{
Example functions for argument "fun":
function(theta,x)
 dgamma(x, shape = theta, log = TRUE)

function(theta,x)
 dcauchy(x, location = theta, log = TRUE)

function(theta,x)
 dbinom(x, 20, prob = 1 / (1 + exp(- theta)), log = TRUE)
}
\examples{
fgam <- function(theta,x) dgamma(x, shape = theta, log = TRUE)
mylogl(x = data2, func = fgam, interval = c(0,100))
}
